+++ Client_App/Commands/AsyncCommands/ExcelExport/ExcelExportFormsAsyncCommand.cs
@@ -126,8 +126,11 @@ public class ExcelExportFormsAsyncCommand : ExcelExportBaseAllAsyncCommand
         var dbReadOnlyPath = Path.Combine(BaseVM.TmpDirectory, BaseVM.DbFileName + ".RAODB");
         try
         {
-            File.Delete(dbReadOnlyPath);
-            File.Copy(Path.Combine(BaseVM.RaoDirectory, BaseVM.DbFileName + ".RAODB"), dbReadOnlyPath);
+            if (!StaticConfiguration.IsFileLocked(dbReadOnlyPath))
+            {
+                File.Delete(dbReadOnlyPath);
+                File.Copy(Path.Combine(BaseVM.RaoDirectory, BaseVM.DbFileName + ".RAODB"), dbReadOnlyPath);
+            }
         }
         catch
         {
