+++ Client_App/Commands/AsyncCommands/ExcelExport/ExcelExportListOfForms1AsyncCommand.cs
@@ -117,11 +117,8 @@ public class ExcelExportListOfForms1AsyncCommand : ExcelBaseAsyncCommand
         }
         catch
         {
-            return;
-        }
-        finally
-        {
             cts.Dispose();
+            return;
         }
 
         var fullPath = result.fullPath;
@@ -131,8 +128,11 @@ public class ExcelExportListOfForms1AsyncCommand : ExcelBaseAsyncCommand
         var dbReadOnlyPath = Path.Combine(BaseVM.TmpDirectory, BaseVM.DbFileName + ".RAODB");
         try
         {
-            File.Delete(dbReadOnlyPath);
-            File.Copy(Path.Combine(BaseVM.RaoDirectory, BaseVM.DbFileName + ".RAODB"), dbReadOnlyPath);
+            if (!StaticConfiguration.IsFileLocked(dbReadOnlyPath))
+            {
+                File.Delete(dbReadOnlyPath);
+                File.Copy(Path.Combine(BaseVM.RaoDirectory, BaseVM.DbFileName + ".RAODB"), dbReadOnlyPath);
+            }
         }
         catch
         {
