+++ Client_App/Commands/AsyncCommands/ExcelExport/ExcelExportListOfForms2AsyncCommand.cs
@@ -123,12 +123,10 @@ public class ExcelExportListOfForms2AsyncCommand : ExcelBaseAsyncCommand
         }
         catch
         {
+            cts.Dispose();
             return;
         }
-        finally
-        {
-            cts.Dispose();
-        }
+
         var fullPath = result.fullPath;
         var openTemp = result.openTemp;
         if (string.IsNullOrEmpty(fullPath)) return;
@@ -136,8 +134,11 @@ public class ExcelExportListOfForms2AsyncCommand : ExcelBaseAsyncCommand
         var dbReadOnlyPath = Path.Combine(BaseVM.TmpDirectory, BaseVM.DbFileName + ".RAODB");
         try
         {
-            File.Delete(dbReadOnlyPath);
-            File.Copy(Path.Combine(BaseVM.RaoDirectory, BaseVM.DbFileName + ".RAODB"), dbReadOnlyPath);
+            if (!StaticConfiguration.IsFileLocked(dbReadOnlyPath))
+            {
+                File.Delete(dbReadOnlyPath);
+                File.Copy(Path.Combine(BaseVM.RaoDirectory, BaseVM.DbFileName + ".RAODB"), dbReadOnlyPath);
+            }
         }
         catch
         {
